🐛Fix Enigma RNG range overflow

+++ Figlotech.Core/Autokryptex/FiRandom.cs
@@ -81,7 +81,7 @@ namespace Figlotech.Core.Autokryptex {
         }
 
         public int Next(int Minimum, int Maximum) {
-            return (Minimum + this.Gen()) % Maximum;
+            return (int)(((long)Minimum + this.Gen()) % Maximum);
         }
     }
 }
